gitwhisperer/python-backend/requirements.txt
openai>=1.0.0
gitpython>=3.1.0
fastapi>=0.100.0
uvicorn>=0.20.0
rich>=13.0.0
python-dotenv>=1.0.0